+++ packages/podlite-toc/src/index.tsx
@@ -1,8 +1,40 @@
 import React from 'react'
-import {Plugin, Location, mkBlock, PodliteDocument, getFromTree, getTextContentFromNode, mkItemBlock, mkTocItem, mkTocList, mkToc, TocList, Toc, getNodeId} from '@podlite/schema'
+import {Plugin, Location, mkBlock, PodliteDocument, getFromTree, getTextContentFromNode, mkItemBlock, mkTocItem, mkTocList, mkToc, TocList, Toc, getNodeId, BlockImage} from '@podlite/schema'
 import { getTocPod, prepareDataForToc } from './helpers';
 import makeAttrs from 'pod6/built/helpers/config'
-
+import { isNamedBlock } from 'pod6/built/helpers/makeTransformer';
+import { PodNode } from '@podlite/schema';
+export const getContentForToc = (node: PodNode): string => {
+    if (typeof node !== "string" && "type" in node) {
+        if ( node.type === 'block') {
+            if (isNamedBlock(node.name)) {
+                const conf = makeAttrs(node, {})
+                const caption = ((conf, nodeName)=>{
+                    if ( conf.exists('caption') ) {
+                        return conf.getFirstValue('caption')
+                    } else
+                    if ( conf.exists('title') ) {
+                        return conf.getFirstValue('title')
+                    } 
+                    return `${nodeName} not have :caption`
+                })(conf, node.name)
+                return caption
+            }
+            if (node.name == 'image') {
+                if ('caption' in node ) {
+                    // @ts-ignore
+                    return node.caption || 'image not have caption'
+                }
+                
+            }
+            return getTextContentFromNode(node);
+        }
+        if (node.type === 'image') {
+            return node.caption || 'image not have caption'
+        }
+      }
+    return 'Not supported toc element';
+  }
 export const plugin:Plugin =({
         toAst: (writer) => (node) => {
             if (typeof node !== "string" && 'type' in node && 'content' in node && node.type === 'block') {
@@ -14,15 +46,23 @@ export const plugin:Plugin =({
                 }
             }
         },
-        toAstAfter:(writer, processor) => {
-            let fulltree
-            writer.on('start', (tree:PodliteDocument) => {
-                // TODO: tree should pass as third argument in pod6 release
-                fulltree = tree
-            })
+        //@ts-ignore
+        toAstAfter:(writer, processor, fulltree) => {
+            // let fulltree
+            // writer.on('start', (tree:PodliteDocument) => {
+            //     // TODO: tree should pass as third argument in pod6 release
+            //     fulltree = tree
+            // })
             return (node,ctx) => {
             const content = getTextContentFromNode(node)
-            const blocks = content.trim().split(/(?:\s*,\s*|\s+)/).filter(Boolean)
+            const blocks = content.trim().split(/(?:\s*,\s*|\s+)/)
+                                    .filter(Boolean)
+                                    .map(blockName=>{
+                                        if (blockName.toLowerCase() == 'image') {
+                                            return 'image'
+                                        }
+                                        return blockName
+                                    })
             const nodes = getFromTree(fulltree, ...blocks)
             const tocTree = prepareDataForToc(nodes)
             const createList = (items:any[], level):TocList=>{
@@ -30,7 +70,7 @@ export const plugin:Plugin =({
                 items.map(item => {
                     const {level, node, content} = item
                     // create new node for each item
-                    const text = getTextContentFromNode(node)
+                    const text = getContentForToc(node)
                     //TODO: getNodeId should use ctx of node, but using {} instead
                     const para = `L<${text}|#${getNodeId(node,{})}>`  
                     const tocNode = processor(para)[0];